requirements.txt
python>=3.10
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
torch>=2.0.0
sentence-transformers>=2.2.0
datasets>=4.8.2
pyyaml>=6.0
matplotlib>=3.6.0
seaborn>=0.12.0
